get all the task list information endpoint functional
